requirements.txt
langchain-core>=0.2.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.1
streamlit>=1.34.0
